modify long to double

+++ doctor-event-api/src/main/java/io/terminus/doctor/event/dto/report/DoctorSaleDailyReport.java
@@ -27,12 +27,12 @@ public class DoctorSaleDailyReport implements Serializable {
     /**
      * 保育猪(产房 + 保育)
      */
-    private long nursery;
+    private double nursery;
 
     /**
      * 育肥猪
      */
-    private long fatten;
+    private double fatten;
 
     public void addBoarSowCount(DoctorSaleDailyReport doctorSaleDailyReport){
         this.boar += doctorSaleDailyReport.getBoar();
